Asset type tests moved to separate class
