+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/UpdateMasterData/AssetTypeTests.cs
@@ -1,0 +1,48 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Threading.Tasks;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Components;
+using Energinet.DataHub.MeteringPoints.IntegrationTests.Tooling;
+using Xunit;
+using Xunit.Categories;
+
+namespace Energinet.DataHub.MeteringPoints.IntegrationTests.UpdateMasterData
+{
+    [IntegrationTest]
+    public class AssetTypeTests : TestHost
+    {
+        public AssetTypeTests(DatabaseFixture databaseFixture)
+            : base(databaseFixture)
+        {
+        }
+
+        [Fact]
+        public async Task Asset_type_is_changed()
+        {
+            await SendCommandAsync(Scenarios.CreateVEProduction()).ConfigureAwait(false);
+
+            var request = CreateUpdateRequest()
+                with
+                {
+                    AssetType = AssetType.Boiler.Name,
+                };
+
+            await SendCommandAsync(request).ConfigureAwait(false);
+
+            AssertMasterData()
+                .HasAssetType(AssetType.Boiler);
+        }
+    }
+}
