+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/UpdateMasterData/ConsumptionMeteringPoints/ChangeTests.cs
@@ -134,23 +134,6 @@ namespace Energinet.DataHub.MeteringPoints.IntegrationTests.UpdateMasterData.Con
         }
 
         [Fact]
-        public async Task Asset_type_is_changed()
-        {
-            await SendCommandAsync(Scenarios.CreateVEProduction()).ConfigureAwait(false);
-
-            var request = CreateUpdateRequest()
-                with
-                {
-                    AssetType = AssetType.Boiler.Name,
-                };
-
-            await SendCommandAsync(request).ConfigureAwait(false);
-
-            AssertMasterData()
-                .HasAssetType(AssetType.Boiler);
-        }
-
-        [Fact]
         public async Task Unit_type_is_changed()
         {
             await SendCommandAsync(Scenarios.CreateVEProduction()).ConfigureAwait(false);
